fixed init bullet's position

+++ src/main/scala/sk/kave/tank/events/Events.scala
@@ -12,7 +12,7 @@ abstract class TankEvent extends Event
 abstract class MapEvent extends Event
 
 case class MapChangeEvent(col: Int, row: Int, newValue: Items) extends MapEvent
-case class ShootEvent(x: Int, y: Int, bullet : Bullet, callback: () => Unit) extends MapEvent {
+case class ShootEvent(x: Double, y: Double, bullet : Bullet, callback: () => Unit) extends MapEvent {
   override def toString:String={
     "ShootEvent: x = " + x + ", y = " + y + ", direction = " + bullet.direction
   }
